+++ app/gui/view/debug_scene/src/lib.rs
@@ -22,6 +22,7 @@
 // === Export ===
 // ==============
 
+pub use debug_scene_cloud_dashboard as cloud_dashboard;
 pub use debug_scene_component_list_panel_view as new_component_list_panel_view;
 pub use debug_scene_icons as icons;
 pub use debug_scene_interface as interface;